+++ src/sort_name.c
@@ -6,7 +6,7 @@
 /*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/05 10:23:24 by iwordes           #+#    #+#             */
-/*   Updated: 2017/01/05 16:19:04 by iwordes          ###   ########.fr       */
+/*   Updated: 2017/01/07 13:50:07 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 ** Sort by ASCII name.
 */
 
-int		sort_name(t_ent *ent1, t_ent *ent2)
+long	sort_name(t_ent *ent1, t_ent *ent2)
 {
 	return (ft_strcmp(ent2->name, ent1->name));
 }
